Removed the (incorrectly named) redundant file graph_list.py and renamed graph.py to graph_list.py (#820)
